Cross two lists at given positions.
Added a number of const to avoid accidental modifications.
New test method that reads elements of lists into an array, instead of
the more cumbersome cumulative sum.

test in makefile builds and runs the test in one command.
Removed warning for unused functions, since help functions such as
display() are often not used.